geometry-mesh/requirements_242.txt
ansys-geometry-core[all]==0.4.14
ansys-meshing-prime==0.6.dev6
